feat: changed ef to npgsql datasource

+++ ConversionSystemReportService/Records/ReportResult.cs
@@ -9,7 +9,7 @@ public record ReportResult
 {
     public sealed record Created(string ReportId) : ReportResult;
 
-    public sealed record Completed(string ReportId, double Conversion, int PurchaseAmount) : ReportResult;
+    public sealed record Completed(string ReportId, double Conversion, long PurchaseAmount) : ReportResult;
     
     public sealed record Failed(string ReportId, string ErrorMessage) : ReportResult;
 };
